fix: address CR changes
