+++ packages/ap-storage/alliance_platform/storage/models.py
@@ -1,0 +1,2 @@
+# this is just here so the model can be detected for migrations
+from alliance_platform.storage.async_uploads.models import AsyncTempFile as AsyncTempFile
